requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pypdf>=3.17.4
docx2txt>=0.8
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.31.0
huggingface_hub>=0.20.0
groq>=0.4.0

